backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.115.12
    # via backend (pyproject.toml)
fastapi-cli==0.0.7
    # via fastapi
google-auth==2.40.0
    # via google-genai
google-genai==1.13.0
    # via backend (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   google-genai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.4
    # via
    #   backend (pyproject.toml)
    #   fastapi
    #   google-genai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
python-dotenv==1.1.0
    # via
    #   backend (pyproject.toml)
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via google-genai
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.5
    # via fastapi-cli
rsa==4.9.1
    # via google-auth
ruff==0.11.8
    # via backend (pyproject.toml)
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
typer==0.15.3
    # via fastapi-cli
typing-extensions==4.13.2
    # via
    #   fastapi
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via
    #   backend (pyproject.toml)
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
